Update README.md
The Logistic Regression Algorithm deals in discrete values whereas the Linear Regression Algorithm handles predictions in continuous values. This means that Logistic Regression is a better option for binary classification. An event in Logistic Regression is classified as 1 if it occurs and it is classified as 0 otherwise.

Logistic regression is a supervised learning classification algorithm used to predict the probability of a target variable. The nature of target or dependent variable is dichotomous, which means there would be only two possible classes.

This assignment will study following Question :
Predicting Customer Response to Telemarketing Campaigns for Term Deposit

Output variable - y - Whether the client has subscribed a term deposit or not

Binomial ("yes" or "no")